devel/bossa: Fix build after recent Clang update on -HEAD
Clang 6 caught a bogus comparison that was not previously caught; fix the
build by correcting the comparison to correctly dereference the `char *`
before comparing it with '\0'.

PORTREVISION bumped because the comparison was erroneous and still
successfully built everywhere but -HEAD.

This patch is effectively from upstream. It will go away with the update to
BOSSA 1.8, which requires a little more testing still.

Approved by:	antoine