Remove the CCACHE option and other ccache code from the Makefile and use the ccache code in the ports framework handle all of the details.
Don't prepend ${WRKSRC}/solenv/bin to PATH in CONFIGURE_ENV since
it is no longer needed and blocks the framework from adding
CCACHE_WRAPPER_PATH to PATH in CONFIGURE_ENV and MAKE_ENV.

Sanitize MAKE_ENV and export it to the build, similar to what is
done by the default do-build target.  This passes CCACHE_DIR and
HOME (set to WRKDIR) to the build, so the $HOME override in r459363
is no longer needed.  We need to remove the CPPFLAGS setting from
MAKE_ENV before exporting to the build because it breaks the build
of the bundled icu module.  CFLAGS and CXXFLAGS are also removed
because they would interfere with the carefully constructed compiler
flags set by the port.  The include paths are already handled by
the port, and the port also modifies the optmization flags for some
files to work around compiler bugs.

While here, remove files/generate.pl since it has been obsolete for
a long while.

No PORTREVISION bump since the package should be unchanged.

PR:		224276


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@459491 35697150-7ecd-e111-bb59-0022644237b5