New port: devel/py-QtPy: Abstraction layer on top of the various Qt bindings
This is an abstraction layer on top of the various Qt bindings.

It doesn't have any explicit dependencies. It has alternative dependencies
on PyQt5, PyQt4, PySide2 or PySide which it selects in the run-time.
Users have the responsibility to ensure that one of them is installed, and
to optionally set QT_API to force one of them.

Approved by:	tcberner (mentor)
Differential Revision:	D13657