graphics/darktable: Include `stddef.h` when testing zError() symbol
The `CheckZlib.cmake` module uses `NULL` in the test source code to
check for the `zError()` symbol. This fails to build on FreeBSD 10.3
with the following error:

    CheckPrototypeDefinition.c:15:10: error: use of undeclared identifier 'NULL'
      return NULL;
             ^

I don't know the root cause for this failure, but including `stddef.h`
in this test source code fixes the problem.

PR:		225501
Reported by:	cpm@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@460168 35697150-7ecd-e111-bb59-0022644237b5