Upgrade to upstream svn revision r1822069, which includes the fixes that were in files/patch-security.
Follow upstream and disable GNOMEVFS by default, using gio instead.

Don't enable crashdumps by default.  This build knob causes the build
to be done with debug symbols, which greatly increase the size of the
build, but this is all in vain because the executables and libraries
are subsequently stripped.

Build with debug symbols if WITH_DEBUG is set.

Build with -fstack-protector.

Only fetch unowinreg.dll when building the SDK.

Remove tomcat from distinfo.  It was leftover from an earlier
experiment.

Add three new options:
dd three new options:

  CRASHDUMP - Enable crashdumps and set WITH_DEBUG.

  DBGUTIL   - Enable assertions, object counting, and other
              non-production runtime debugging.

  DEBUG     - Compile with -O0 for better debugability.  The ports
              framework also sets WITH_DEBUG.