PyQt: Unconditionally install the .pyi files
r460809 started installing .pyi with the %%PYTHON3%% substitution in the
plists. Support for .pyi files was added in Python 3.5, and PyQt's configure.py
only installs the files on Python >= 3.5.

This patch removes the version checks from configure.py (the files are just
unused in earlier Python versions), as it is easier than checking Python 3's
version in each PyQt port's Makefile (or to add the logic to Mk/Uses/pyqt.mk).

PR:		225773
Reviewed by:	tcberner
Differential Revision:	https://reviews.freebsd.org/D14361


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@461841 35697150-7ecd-e111-bb59-0022644237b5