Update QScintilla 2.10.2.
Upstream has renamed its libraries and we no longer need to patch
devel/qscintilla-qt5 to create a library whose name does not conflict with
devel/qscintilla's. However, the library names are different so we need to bump
PORTREVISION in several ports (the SOVERSION has changed too).

Thanks to antoine for the exp-run.

PR:		225928