new port: graphics/drm-stable-kmod
This is a repo copy of the last 4.9 status of graphics/drm-next-kmod. While drm-next-kmod serves as the experimental DRM/KMS port, this new port will always trail it and be in a more stable state. This is not a guarantee that drm-stable-kmod will receive all versions from drm-next-kmod, more likely there will be jumps.

Mark the two ports as mutually conflicting.

Reviewed by:	swills (mentor)
Approved by:	swills (mentor)
Differential Revision:	https://reviews.freebsd.org/D14584


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@464646 35697150-7ecd-e111-bb59-0022644237b5