update sysutils/vagrant to 2.0.3
The port's upstream github account was changed/renamed, this is reflected
in the GH_ACCOUNT change.
Tests are not installed anymore, so pkg-plist is a lot shorter now.
The ".runner" "simple build script" is not required for the port and
installed by accident, so just rm it.

PR:		226662
Approved by:	joe at thrallingpenguin com


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@464729 35697150-7ecd-e111-bb59-0022644237b5