Update to a newer snapshot from it's source. Move to the official github mirror, since this allows to use the USE_GITHUB helper. Unfortunately this snapshot is not tagged so we have to use GH_TAGNAME.
One big advantage is that this new version does not depend on the deprecated
x11-toolkits/libxfce4gui component anymore.

Since it's a in development plugin it does not provide a configure script but
depends on devel/xfce4-dev-tools to generate the required parts.

Move the two files from pkg-plist to PLIST_FILES.

PR:		226821
Submitted by:	madpilot