audio/oss: set LLD_UNSAFE to link using ld.bfd
This port fails when attempting to link with lld as the system linker.
It passes -fPIC to the linker, which is actually a compiler flag.
ld.bfd accepts the flag but it results in a bogus DT_AUXILIARY entry
in the ouptut:

Dynamic section at offset 0x5000 contains 15 entries:
  Tag                Type                  Name/Value
 0x000000007ffffffd AUXILIARY            Auxiliary library: [PIC]

PR:		221807
Approved by:	portmgr (LLD_UNSAFE blanket)
Sponsored by:	The FreeBSD Foundation