Remove ElasticSearch X-Pack packages
This is not a supported way to install the X-Pack plugins. The software
comes with tools for installing and managing plugins. The plugins are a
zip file with a payload that configures various system settings and a
file/directory structure that does not match what these packages
provide. The end result is a partially functional plugin with a myriad
of issues due to files and configs missing from expected locations.

Please use the elasticsearch-plugin and kibana-plugin commands to
install X-Pack.