Deprecate FreeMAT.
Set the expiration date to just after the 5th anniversary of the last
release (this should allow it to appear in the 2018Q3 quarterly to allow
more time for mirgration).

FreeMAT is large and complex software with no sign of upstream
development (or even list/group traffic).  It is the sole remaining
dependency on lang/clang33 and devel/llvm33.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@469232 35697150-7ecd-e111-bb59-0022644237b5