Add a new port: sysutils/sockaddr
Sockaddr is a command-line utility that makes it easy to query network
address information from the local server. With sockaddr it is possible
to create a list of the RFC 1918 addresses on a server, obtain the
default gateway, or public IP addresses and use this information in other
shell scripts. See the website for additional templating and IP math
functions.

Reviewed by:	mat, swills
Approved by:	swills (mentor)
Differential Revision:	https://reviews.freebsd.org/D14570


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@469691 35697150-7ecd-e111-bb59-0022644237b5