multimedia/motion:
- Add WEBCAMD option and enable by default
- Bump PORTREVISION

From the initiating PR:

The multimedia/pwcbsd port contains an kernel module that appears to no longer
work.  The multimedia/pwcbsd port distfile has not been updated since 2014 so
it is very unlikely to work.  The multimedia/pwcbsd port was long ago replaced
by the multimedia/webcamd port.

The current multimedia/motion port has a PWCBSD option.  This sets
--with-pwcbsd in the configure step, create a build depend on v4l_compat and
creates a run depend on multimedia/pwcbsd, which is then built and installed
but not used.  The motion application reads from /dev/video* which is now
handled by multimedia/webcamd.

The patch provided adds an option named WEBCAMD to an existing radio option
that contains the PWCBSD option.  The WEBCAMD option sets --with-pwcbsd in the
configure step, creates the build depend on v4l_compat but does not create a
run depend on pwcbsd.  It could create a run depend on multimedia/webcamd but I
didn't add that.

The only affect of --with-pwcbsd in the configure step of the downstream port
is to suppress disabling video4linux support and link with v4l_compat library.
At some point the downstream should be advised to change the option names in
the configure.ac file to reflect the demise of pwcbsd.

I left the PWCBSD option in place and unchanged.  It may serve no useful
purpose so it could be removed.

PR:		228102
Submitted by:	Curtis Villamizar <curtis@ipv6.occnc.com>
MFH:		yes