Fix checking of torrents which contain files larger than 2^31 - 1 bytes.
By default it detects file size by fseek()'ing to its end and ftell()'ing
current position, and these functions take `long offset' as an argument.
It is sufficient on 64-bit architectures where sizeof(long) is 8, but not
so on 32-bit ones where it is 4.

Make it work everywhere by building with -DUSE_FTELLO so it uses fseeko(3)
and ftello(3) functions which take `off_t offset'.

Tested on:	i386, powerpc