x11-servers/xorg-server: Backport security fixes
Backport security fixes for CVE-2017-10971 and CVE-2017-10972 (yes, 2017).
For some reason this was not done when the vulnerabilities were documented
in VuXML, and a typo in the version range in VuXML meant that the entries
never matched.

This fixes a memory disclosure and a couple of buffer overruns.

PR:		220584
Reported by:	Vladimir Krstulja
MFH:		2018Q2
Security:	ab881a74-c016-4e6d-9f7d-68c8e7cedafb