Fix building after r470315
When I switched from autoreconf, I forgot to point this patch file at
Makefile.in instead of Makefile.am. On some systems, this caused
automake to be run because it saw the Makefile.am as newer than the
Makefile.in, but automake isn't pulled in as a dependency.

Reported by:	antoine