Fix unicode for 'C' locale.
wxPython 3.0 uses the current locale when wx is imported to decide on the
default text encoding.  When the locale is 'C' this means only US-ASCII is
supported, corrupting the disaply quite badly.  As a hack, if the encoding
is US-ASCII, force LC_ALL to en_US.UTF-8 which allows unicode to work.

Reported by:	db