math/libflame: bring to current date and take maintainership
This brings our math/libflame to a recent state, as opposed to the outdated one before. Their development has moved to
github. Configure it as a LAPACK library and to use an external BLAS to eventually support a blaslapack:flame together
with math/blis. Please note that this should be used with caution in production until the port is better tested!

Their make system now generates a large quantity of object files to link which I found to only work with a recent
linker (hence the requirement for recent CURRENT with lld default). Since this is a numerical library, some caution is
required before we can enable it for other architectures. Patches were generated by hand as makepatch failed to work
properly.

Reviewed by:	mat, zeising (mentor)
Approved by:	zeising (mentor)
Differential Revision:	https://reviews.freebsd.org/D15716


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@473932 35697150-7ecd-e111-bb59-0022644237b5