Sphinx Extensions to support API reference sites in OpenStack.
The package is a collection of sphinx stanzas that assist in building an API
Reference site for an OpenStack project in RST. RST is great for unstructured
English, but displaying semi structured (and repetitive) data in tables is not
its strength. This provides tooling to insert semi-structured data describing
request and response parameters and status or error messages, and turn those
into nice tables.

The package also includes a set of styling (and javascript) that is expected to
layer on top of a Sphinx theme base. This addition provides a nice set of
collapsing sections for REST methods and javascript controls to expand or
collapse all sections.

WWW: https://docs.openstack.org/os-api-ref/

PR:		228846
Submitted by:	freebsd_ports@k-worx.org
Sponsored by:	iXsystems Inc.