Enable Bash completion.
Arcanist is shipped with a Bash completion script. The script isn't installed
in the directories of the bash-completion library so it's never activated.

This patch creates a symbolic link to the script in the directories of the
bash-completion library so that the library is able to pick up the completion
script for arc when a user requests a command completion for the first time.

Reviewed by:	grembo (maintainer), mat (mentor)
Approved by:	mat (mentor)
Differential Revision:	https://reviews.freebsd.org/D16345


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@474969 35697150-7ecd-e111-bb59-0022644237b5