mail/qmail-dk: Fix build with lld
/usr/bin/ld: error: duplicate symbol: scan_ulong
>>> defined at scan_ulong.c
>>>            scan_ulong.o:(scan_ulong) in archive fs.a
>>> defined at scan_ulong.c
>>>            scan_ulong.o:(.text+0x0)

http://beefy12.nyi.freebsd.org/data/head-amd64-default/p474767_s336359/logs/qmail-dk-0.54_1.log

scan_ulong.o is added to fs.a and appears twice in the link command
line (via fs.a and listed explicitly) which results in a duplicate
symbol error with lld.  Removing scan_ulong.o from the link command
line would result in an undefined reference error when linking
auto-int with ld.bfd, so remove it from fs.a instead.

Attempt to build on aarch64 again as the error message is the same.

PR:		221812
Reported by:	emaste, pkg-fallout