New port: graphics/p5-Image-PNG-QRCode
make a PNG image containing a QR code from text

This module converts input text to a PNG image of a QR code containing the
text. The PNG image can either be stored to a file or it can be a scalar.

Image::PNG::Libpng requires "libpng" to be installed. "libpng" should
already be installed on most Linux and Windows systems.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@475449 35697150-7ecd-e111-bb59-0022644237b5