Upgrade from 4.0.10 to 4.0.11.
<ChangeLog>

Upgrade urgency HIGH: not critical but very important bugs fixed.

Dear users, this is just a bugfix release of Redis 4. All new work
is now focused on Redis 5, however we backported a number of bug fixes here:

* The disconnection time between the master and slave was reset in an
  incorrect place, sometimes a good slave will not be able to failover
  because it claims it was disconnected for too much time from the master.
* A replication bug, rare to trigger but non impossible, is in Redis for
  years. It was lately discovered at Redis Labs and fixed by Oran Agra.
  It may cause disconnections, desynchronizations and other issues.
* RANDOMKEY may go in infinite loop on rare situations. Now fixed.
* EXISTS now works in a more consistent way on slaves.
* Sentinel: backport of an option to deny a potential security problem
  when the SENTINEL command is used to configure an arbitrary script
  to execute.

Many of these issues are there for a very long time, however upgrading
is a good idea.

</ChangeLog>