Update unixcw and xcwcp to latest release.
This is also the first version where all the issues are fixed, and support
has been included for ALSA and Pulseaudio.  The ioctl() issue is resolved,
and gettext is properly controlled by the NLS option.

I'll be submitting these patches upstream, but due to the low rate of
releases, want to get them into ports first.

This is heavily modified from the submitted patch as every time anything
was tested, a new issue cropped up.

PR:		229782
Reported by:	Yuri Victorovich <yuri@freebsd.org>