net/omnitty: Add abililty to "toggle" tagged machines
We often need to do "something" to about half the machines, then
do "something else" to the other half.  So add a hack to toggle the
"tagged" status of the machines.

This is implemented by means of the F1 menu; the command characters
"o" and "O" can toggle the "tagged" status of the machines.

PR:		211682
Submitted by:	maintainer