Fix build when the default python version isn't 2.7.
The problem here is that if the default version was (e.g.) 3.6, a
python 3.6 version of sphinx would be installed, but a python 2.7
version of py-recommonmark would be installed and then not be found.
Fix this problem by forcing the install of a python 2.7 version of
sphinx.

No functional change.