Fix build when DEFAULT_VERSIONS=python=3.[4-6] is set
In a python 3.x environment, the unversioned symlink to sphinx-build points
to the python 3.x version which will cause the build to fail. This port
can only use python 2.7, so we need to specify the version of sphinx-build.

This was supposedly fixed in r467840 and the PR was closed, but I was
informed by fluffy via private mail that it was still broken.

PR:		227260
Submitted by:	fluffy
Approved by:	portmgr (blanket - trivial build fix)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@478216 35697150-7ecd-e111-bb59-0022644237b5