Upgrade from 5.0-rc4 to 5.0-rc5.
<ChangeLog>

Upgrade urgency HIGH: Several imporant bugs fixed.

This is the release candidate number five, and has a lot of bug fixes inside,
together with a few big changes to the Redis behavior from the point of view
of replication of scripts and handling of the maxmemory directive in slaves.
Make sure to read the whole list!

* BREAKING BEHAVIOR: Slaves now ignore maxmemory by default.
* BREAKING BEHAVIOR: Now scripts are always replicated for their effects, and
                     never sending the script itself to slaves/AOF.
* Improvement: Big pipelining performances improved significantly.
* Fix: Rewrite BRPOPLPUSH as RPOPLPUSH to propagate.
* Fix: False positives in tests.
* Fix: Certain command flags were modified because not correct.
* Fix: Fix blocking XREAD for streams that are empty.
* Improvement: Allow scripts to timeout on slaves as well.
* Fix: Different corner cases due to CLIENT PAUSE are now fixed.
* Improvement: Optimize parsing large bulk greater than 32k.
* Fix: Propagate read-only scripts as SCRIPT LOAD, not as EVAL.

</ChangeLog>