devel/flang[-clang]: remove wrapper script dependency.
Patch flang-clang directly to link the required libraries and include the proper
library directories. Hence, the need of the wrapper script (and it's non-optimal
behavior) vanishes.

While there, mark flang-clang broken on 10. Its LLVM6 base is not supported
there and due to libpgmath only working on 12+, there is no need for it
anyways.