Make X-Face support optional and disable it by default
X-Face support is disabled in the default configuration of muttprint, and the
support for this option requires mail/faces which in turn requires devel/glib12
and x11-toolkits/gtk12.  Make it possible for users who really care about this
feature to get it, and remove these old dependencies for others.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@479586 35697150-7ecd-e111-bb59-0022644237b5